conversao-planilhas/requirements.txt
fastapi[all]
sqlalchemy
asyncmy
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-multipart
bcrypt==3.2.2
unidecode
openpyxl